+++ gusto/src/main/java/com/umc/gusto/domain/user/controller/UserController.java
@@ -1,5 +1,6 @@
-package com.umc.gusto.domain.user;
+package com.umc.gusto.domain.user.controller;
 
+import com.umc.gusto.domain.user.service.UserService;
 import com.umc.gusto.domain.user.model.request.SignUpRequest;
 import com.umc.gusto.global.auth.model.Tokens;
 import lombok.RequiredArgsConstructor;
