+++ gusto/src/main/java/com/umc/gusto/domain/user/model/request/SignUpRequest.java
@@ -10,6 +10,4 @@ public class SignUpRequest {
     String age;
     String gender;
     String profileImg;
-
-
 }
